+++ java-desktop-util-chart/src/main/java/ec/util/chart/ObsFunction.java
@@ -17,7 +17,7 @@
 package ec.util.chart;
 
 import lombok.NonNull;
-import org.checkerframework.checker.nullness.qual.Nullable;
+import org.jspecify.annotations.Nullable;
 
 import java.util.Locale;
 import java.util.Objects;