+++ java-desktop-util-os/src/main/java/ec/util/desktop/impl/XdgDesktop.java
@@ -25,7 +25,7 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 import nbbrd.service.ServiceProvider;
 import lombok.NonNull;
-import org.checkerframework.checker.nullness.qual.Nullable;
+import org.jspecify.annotations.Nullable;
 
 /**
  * A generic {@link Desktop} implementation for Linux.<p>