+++ java-desktop-util-os/src/test/java/ec/util/desktop/impl/MapSystem.java
@@ -1,7 +1,7 @@
 package ec.util.desktop.impl;
 
 import lombok.NonNull;
-import org.checkerframework.checker.nullness.qual.Nullable;
+import org.jspecify.annotations.Nullable;
 
 import java.io.IOException;
 import java.util.Map;