+++ java-desktop-util-swing/src/main/java/ec/util/various/swing/BasicSwingLauncher.java
@@ -17,7 +17,7 @@
 package ec.util.various.swing;
 
 import lombok.NonNull;
-import org.checkerframework.checker.nullness.qual.Nullable;
+import org.jspecify.annotations.Nullable;
 
 import javax.swing.*;
 import java.awt.*;